MacroNode: formatting fix + refactored away some compiler warninings (signed/unsigned comparisons)
OctileHeuristic: better output when crashing

DebugUtility: doesn't crash if we pass a null graph object for printing

+++ trunk/src/rsr/MacroNode.cpp
@@ -15,24 +15,27 @@ MacroNode::~MacroNode()
 
 }
 
-void MacroNode::reset()
+void 
+MacroNode::reset()
 {
 	ClusterNode::reset();
 	setParent(0);
 }
 
-void MacroNode::addSecondaryEdge(edge* e)
+void 
+MacroNode::addSecondaryEdge(edge* e)
 {
 	secondaryEdges.push_back(e); 
 }
 
-void MacroNode::removeSecondaryEdge(int edgeNum)
+void 
+MacroNode::removeSecondaryEdge(unsigned int edgeNum)
 {
 	std::vector<edge*>::iterator iter = secondaryEdges.begin();
 	while(iter != secondaryEdges.end())
 	{
 		edge* e = *iter; 
-		if(e->getEdgeNum() == edgeNum)
+		if(e->getEdgeNum() == (int)edgeNum)
 		{
 			secondaryEdges.erase(iter);
 			break;
@@ -41,7 +44,8 @@ void MacroNode::removeSecondaryEdge(int edgeNum)
 	}
 }
 
-edge* MacroNode::getSecondaryEdge(int index)
+edge* 
+MacroNode::getSecondaryEdge(unsigned int index)
 {
 	if(0 <= index && index <= secondaryEdges.size())
 		return secondaryEdges.at(index);
